refactor(api): move user route tests in its own directory
