+++ api/tests/identity-access-management/integration/application/user/user.route.test.js
@@ -1,6 +1,6 @@
-import { identityAccessManagementRoutes } from '../../../../src/identity-access-management/application/routes.js';
-import * as i18nPlugin from '../../../../src/shared/infrastructure/plugins/i18n.js';
-import { expect, HttpTestServer } from '../../../test-helper.js';
+import { identityAccessManagementRoutes } from '../../../../../src/identity-access-management/application/routes.js';
+import * as i18nPlugin from '../../../../../src/shared/infrastructure/plugins/i18n.js';
+import { expect, HttpTestServer } from '../../../../test-helper.js';
 
 const routesUnderTest = identityAccessManagementRoutes[0];
