+++ api/tests/identity-access-management/unit/application/user/user.route.test.js
@@ -1,8 +1,8 @@
-import { identityAccessManagementRoutes } from '../../../../src/identity-access-management/application/routes.js';
-import { userController } from '../../../../src/identity-access-management/application/user/user.controller.js';
-import { userVerification } from '../../../../src/identity-access-management/application/user/user-existence-verification-pre-handler.js';
-import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
+import { identityAccessManagementRoutes } from '../../../../../src/identity-access-management/application/routes.js';
+import { userController } from '../../../../../src/identity-access-management/application/user/user.controller.js';
+import { userVerification } from '../../../../../src/identity-access-management/application/user/user-existence-verification-pre-handler.js';
+import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
+import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 const routesUnderTest = identityAccessManagementRoutes[0];
